pyproject.toml
[project]
name = "sentient"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu",
    "langchain<1.0.0",
    "langchain-community",
    "langchain-huggingface",
    "langchain-openai",
    "pypdf",
    "python-dotenv",
    "streamlit",
]
